Move the tests to the proper places.
