+++ candle-core/src/tensor.rs
@@ -2742,98 +2742,3 @@ impl std::ops::Div<&Tensor> for f64 {
         rhs.recip()? * self
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::{DType, Device, Tensor};
-    use anyhow::Result;
-    fn assert_tensor_eq(t1: &Tensor, t2: &Tensor) -> Result<()> {
-        assert_eq!(t1.shape(), t2.shape());
-        // Default U8 may not be large enough to hold the sum (`t.sum_all` defaults to the dtype of `t`)
-        let eq_tensor = t1.eq(t2)?.to_dtype(DType::U32)?;
-        let all_equal = eq_tensor.sum_all()?;
-        assert_eq!(all_equal.to_scalar::<u32>()?, eq_tensor.elem_count() as u32);
-        Ok(())
-    }
-
-    #[test]
-    fn test_flip_1d() -> Result<()> {
-        // 1D: [0, 1, 2, 3, 4]
-        let t = Tensor::arange(0.0, 5.0, &Device::Cpu)?.reshape((5,))?;
-        let flipped = t.flip(&[0])?;
-        // Expected: [4, 3, 2, 1, 0]
-        let expected = Tensor::from_vec(vec![4.0, 3.0, 2.0, 1.0, 0.0], (5,), &Device::Cpu)?;
-        assert_tensor_eq(&flipped, &expected)?;
-        Ok(())
-    }
-
-    #[test]
-    fn test_flip_2d() -> Result<()> {
-        // 2D:
-        // [[0, 1, 2],
-        //  [3, 4, 5]]
-        let t = Tensor::arange(0.0, 6.0, &Device::Cpu)?.reshape((2, 3))?;
-        let flipped = t.flip(&[0, 1])?;
-        // Expected:
-        // [[5, 4, 3],
-        //  [2, 1, 0]]
-        let expected = Tensor::from_vec(vec![5.0, 4.0, 3.0, 2.0, 1.0, 0.0], (2, 3), &Device::Cpu)?;
-        assert_tensor_eq(&flipped, &expected)?;
-        Ok(())
-    }
-
-    #[test]
-    fn test_flip_3d_channels() -> Result<()> {
-        // 3D:
-        // [[[0,1,2],
-        //   [3,4,5]],
-        //
-        //  [[6,7,8],
-        //   [9,10,11]]]
-        let t = Tensor::arange(0.0, 12.0, &Device::Cpu)?.reshape((2, 2, 3))?;
-        let flipped = t.flip(&[2])?;
-        // Expected:
-        // [[[2,1,0],
-        //   [5,4,3]],
-        //
-        //  [[8,7,6],
-        //   [11,10,9]]]
-        let expected = Tensor::from_vec(
-            vec![2.0, 1.0, 0.0, 5.0, 4.0, 3.0, 8.0, 7.0, 6.0, 11.0, 10.0, 9.0],
-            (2, 2, 3),
-            &Device::Cpu,
-        )?;
-        assert_tensor_eq(&flipped, &expected)?;
-        Ok(())
-    }
-
-    #[test]
-    fn test_flip_backprop() -> Result<()> {
-        let device = &Device::Cpu;
-
-        // Create a tensor (leaf node) that requires gradients
-        let x = Tensor::ones((2, 2), DType::F64, device)?.make_var()?;
-        let weights = Tensor::arange(1.0, 5.0, device)?.reshape((2, 2))?;
-
-        let y = x.matmul(&weights)?;
-        let expected_y = Tensor::from_vec(vec![4.0, 6.0, 4.0, 6.0], (2, 2), device)?;
-        assert_tensor_eq(&y, &expected_y)?;
-
-        let z = y.flip(&[1])?;
-        let expected_z = Tensor::from_vec(vec![6.0, 4.0, 6.0, 4.0], (2, 2), device)?;
-        assert_tensor_eq(&z, &expected_z)?;
-
-        let loss = z.sum_all()?;
-
-        let grad_store = loss.backward()?;
-        let grad_x = grad_store.get_id(x.id()).unwrap();
-
-        let flipped_weights = weights.flip(&[1])?;
-        let dloss_dy = Tensor::ones((2, 2), DType::F64, device)?;
-        // dloss/dx = dloss/dy @ dy/dx = ones @ weight.flip.T
-        let expected_grad = dloss_dy.matmul(&flipped_weights.t()?)?;
-        assert_tensor_eq(&grad_x, &expected_grad)?;
-
-        Ok(())
-    }
-}
